// src/interfaces/collaborator.interface.ts
export interface ICollaborator {
    id: number;
    permissao: number;
    nome: string;
    centro_custo: string;
    data_nascimento: Date;
    sexo: number;
    rg: string;
    orgao_expedidor: string;
    uf_rg: string;
    cpf: string;
    pis_pasep: string;
    ctps: string;
    exped_ctps: Date;
    serie_ctps: string;
    uf_ctps: string;
    email: string;
    setor: string;
    logradouro: string;
    bairro: string;
    cidade: string;
    uf: string;
    cep: string;
    telefone: string;
    naturalidade: string;
    nome_mae: string;
    nome_pai: string;
    unidade: string;
    id_empresa: number;
    nome_empresa: string;
    id_cargo: number;
    nome_cargo: string;
    matricula: string;
    codigo_interno: string;
    situacao: number;
    data_admissao: Date;
    data_rescisao: Date;
    descricao: string;
    etnia: number;
    pcd: boolean;
    tipo_pcd: string;
    cbo: string;
    ccbo_2002: string;
    codigo_filial: string;
    nome_filial: string;
    bairro_filial: string;
    estado_filial: string;
    cep_filial: string;
    cnpj_filial: string;
    telefone_filial: string;
    senha?: string;
}